Refactor table and image io functions
